Move vec_string related stuff to new file
